@@ -1,6 +1,7 @@
 {
   "name": "podcastdx-client",
   "version": "5.1.0",
+  "type": "module",
   "main": "dist/src/index.js",
   "types": "dist/src/index.d.ts",
   "license": "MIT",
@@ -19,39 +20,40 @@
     "test": "jest",
     "lint": "eslint --ext .js,.ts src/",
     "test:watch": "jest --watch",
-    "test:episodes": "DEBUG=* jest episodes",
-    "test:podcasts": "DEBUG=* jest podcasts",
-    "test:recent": "DEBUG=* jest recent",
-    "test:search": "DEBUG=* jest search",
-    "validate": "DEBUG=* ts-node -r tsconfig-paths/register src/schemas/validate.ts",
-    "dev": "DEBUG=* ts-node -r tsconfig-paths/register example.ts",
-    "dev:watch": "ts-node-dev --respawn --no-notify example.ts"
+    "dev": "cross-env DEBUG=* ts-node -r tsconfig-paths/register example.ts",
+    "dev:watch": "cross-env DEBUG=* ts-node-dev --respawn --no-notify example.ts",
+    "test:episodes": "cross-env DEBUG=* jest episodes",
+    "test:podcasts": "cross-env DEBUG=* jest podcasts",
+    "test:recent": "cross-env DEBUG=* jest recent",
+    "test:search": "cross-env DEBUG=* jest search",
+    "validate": "cross-env DEBUG=* ts-node -r tsconfig-paths/register src/schemas/validate.ts"
   },
   "devDependencies": {
     "@types/debug": "^4.1.5",
-    "@types/jest": "^26.0.20",
-    "@types/node": "^14.14.27",
+    "@types/jest": "^30.0.0",
+    "@types/node": "^25.0.10",
     "@types/node-fetch": "^2.5.8",
-    "@types/ramda": "^0.27.38",
-    "@typescript-eslint/eslint-plugin": "^4.1.0",
-    "@typescript-eslint/parser": "^4.1.0",
+    "@types/ramda": "^0.31.1",
+    "@typescript-eslint/eslint-plugin": "^8.53.1",
+    "@typescript-eslint/parser": "^8.53.1",
     "auto-changelog": "^2.2.1",
-    "eslint": "^7.9.0",
-    "eslint-config-airbnb-typescript": "^10.0.0",
+    "cross-env": "^10.1.0",
+    "eslint": "^9.39.2",
+    "eslint-config-airbnb-typescript": "^16.1.1",
     "eslint-config-prettier": "^6.11.0",
-    "eslint-plugin-import": "^2.22.0",
+    "eslint-plugin-import": "^2.32.0",
     "eslint-plugin-prettier": "^3.1.4",
-    "eslint-plugin-sonarjs": "^0.5.0",
+    "eslint-plugin-sonarjs": "^3.0.5",
     "husky": "^5.1.2",
-    "jest": "^26.6.3",
+    "jest": "^30.2.0",
     "lint-staged": "^10.5.4",
     "prettier": "^2.2.1",
     "rimraf": "^3.0.2",
-    "ts-jest": "^26.5.2",
+    "ts-jest": "^29.4.6",
     "ts-node-dev": "1.1.6",
     "tsconfig-paths": "^3.9.0",
-    "typedoc": "^0.20.28",
-    "typescript": "^4.2.2"
+    "typedoc": "^0.28.16",
+    "typescript": "^5.9.3"
   },
   "dependencies": {
     "ajv": "^7.1.1",
